refactor: HPCalendarContentView, HPCalendarDelegateProxy, HPCalendarInterface, HPCalendarProxyBinder Folder Group 추가 - HPCalendarBubbleDayCell State alpha -> color 네이밍 수정
