+++ UI/HPCommonUI/Sources/Common/Calendar/HPCalendarView.swift
@@ -14,11 +14,6 @@ import Then
 import SnapKit
 
 
-public enum CalendarStyle {
-    case `default`
-    case bubble
-}
-
 public final class HPCalendarView: UIView {
     
     public var isStyle: CalendarStyle  {
@@ -221,17 +216,16 @@ public final class HPCalendarView: UIView {
             layoutSize: bubbleCalendarItemSize
         )
         
-        bubbleCalendarLayoutItem.contentInsets = NSDirectionalEdgeInsets(top: 0, leading: 8, bottom: 0, trailing: 0)
+        bubbleCalendarLayoutItem.contentInsets = NSDirectionalEdgeInsets(top: 0, leading: 7, bottom: 0, trailing: 7)
         
         let bubbleCalendarGroup = NSCollectionLayoutGroup.horizontal(
             layoutSize: bubbleCalendarGroupSize,
             subitem: bubbleCalendarLayoutItem,
-            count: 7
+            count: 6
         )
         
         let bubbleCalendarSection = NSCollectionLayoutSection(group: bubbleCalendarGroup)
-        
-        bubbleCalendarSection.contentInsets = NSDirectionalEdgeInsets(top: 0, leading: -40, bottom: 0, trailing: -40)
+        bubbleCalendarSection.contentInsets = NSDirectionalEdgeInsets(top: 0, leading: 16, bottom: 0, trailing: 16)
         bubbleCalendarSection.orthogonalScrollingBehavior = .continuous
         
         
@@ -305,71 +299,4 @@ extension HPCalendarView: ReactorKit.View {
             }).disposed(by: disposeBag)
         
     }
-}
-
-
-
-public final class HPCalendarContentView: UIView {
-    
-    private var disposeBag: DisposeBag = DisposeBag()
-        
-    public lazy var calendarMonthLabel: UILabel = UILabel().then {
-        $0.text = "\(Date().month)월"
-        $0.font = HPCommonUIFontFamily.Pretendard.bold.font(size: 16)
-        $0.textAlignment = .center
-        $0.numberOfLines = 1
-    }
-    
-    public lazy var previousButton: UIButton = UIButton(type: .custom).then {
-        $0.setTitle("", for: .normal)
-        $0.setImage(HPCommonUIAsset.previousArrow.image, for: .normal)
-    }
-    
-    public lazy var nextButton: UIButton = UIButton(type: .custom).then {
-        $0.setTitle("", for: .normal)
-        $0.setImage(HPCommonUIAsset.nextArrow.image, for: .normal)
-    }
-    
-    
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        configure()
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    
-    private func configure() {
-        
-        [calendarMonthLabel, previousButton, nextButton].forEach {
-            self.addSubview($0)
-        }
-        
-        
-        calendarMonthLabel.snp.makeConstraints {
-            $0.width.lessThanOrEqualTo(45)
-            $0.height.equalTo(14)
-            $0.top.equalToSuperview().offset(20)
-            $0.centerX.equalToSuperview()
-        }
-        
-        previousButton.snp.makeConstraints {
-            $0.width.equalTo(16)
-            $0.height.equalTo(17)
-            $0.right.equalTo(calendarMonthLabel.snp.left).offset(-20)
-            $0.centerY.equalTo(calendarMonthLabel)
-        }
-        
-        
-        nextButton.snp.makeConstraints {
-            $0.width.equalTo(16)
-            $0.height.equalTo(17)
-            $0.left.equalTo(calendarMonthLabel.snp.right).offset(20)
-            $0.centerY.equalTo(calendarMonthLabel)
-        }
-    }
-    
 }
